+++ 2013/cat4/pizza/voorbeeld.invoer
@@ -1,0 +1,13 @@
+12
+4 2 3 1 2
+5 1 1 3 6 2
+5 8 8 7 4 13
+7 9 9 1 12 1 4 5
+7 8 14 8 13 3 14 9
+5 13 15 2 14 12
+10 3 13 13 12 1 15 6 2 5 2
+10 13 7 9 8 9 12 14 9 15 6
+12 12 2 4 1 6 11 11 12 1 10 2 10
+20 1 1 3 6 2 2 8 8 7 4 13 14 9 9 1 12 1 4 5 14
+22 14 3 4 2 3 15 11 13 10 5 12 13 11 10 6 10 7 13 12 5 1 10
+27 7 3 7 8 1 10 12 3 8 14 5 15 6 14 14 6 8 11 15 12 12 2 13 3 1 7 5
