+++ 2013/cat4/pizza/voorbeeld.uitvoer
@@ -1,0 +1,12 @@
+5
+9
+25
+26
+39
+29
+47
+60
+49
+77
+103
+113
