+++ 2013/cat4/pizza/wedstrijd.invoer
@@ -1,0 +1,101 @@
+100
+5 7 8 8 15 10
+8 12 11 15 1 4 1 2 3
+5 12 15 2 9 2
+7 1 4 7 15 6 1 15
+8 13 7 14 1 4 7 15 9
+7 8 7 12 5 7 2 3
+7 3 14 11 5 3 13 7
+7 15 8 14 11 4 7 4
+8 1 15 14 2 8 2 12 14
+5 14 1 4 8 10
+5 13 15 2 14 12
+7 8 14 8 13 3 14 9
+7 9 9 1 12 1 4 5
+5 8 8 7 4 13
+5 1 1 3 6 2
+4 2 3 1 2
+7 14 6 8 3 3 6 10
+7 4 4 8 11 1 7 14
+7 15 15 14 8 8 2 8
+7 5 14 12 12 3 4 9
+7 4 13 8 1 15 3 10
+7 13 5 4 3 1 7 8
+7 6 6 13 9 8 15 14
+7 2 10 10 10 10 11 13
+7 4 8 2 14 13 12 7
+7 7 9 11 5 11 3 15
+7 8 9 4 3 3 9 14
+7 12 9 5 9 13 12 7
+7 12 10 7 12 15 5 5
+8 4 6 10 15 7 10 14 7
+8 6 13 4 4 9 6 14 14
+8 3 15 5 11 6 4 15 12
+8 14 10 9 13 4 9 8 3
+8 13 5 8 11 3 6 5 12
+8 11 13 9 8 15 5 5 2
+8 6 1 3 5 9 1 6 15
+8 2 11 13 14 3 12 15 14
+8 7 12 9 13 2 9 12 1
+8 8 3 6 6 3 12 3 11
+8 2 13 15 4 2 8 13 3
+8 1 1 2 9 14 3 14 9
+8 1 7 4 4 13 9 15 1
+8 1 11 5 12 6 1 9 7
+8 12 12 3 13 10 7 4 14
+8 5 2 3 2 4 14 8 13
+8 4 3 7 1 10 5 13 4
+8 3 10 1 6 12 12 2 4
+8 3 15 2 13 15 4 12 9
+8 13 2 5 10 14 4 10 12
+8 4 9 3 3 1 14 4 14
+8 9 2 12 6 11 2 10 13
+8 5 2 9 7 8 6 6 13
+9 7 12 14 11 12 12 8 6 10
+9 15 12 5 4 13 9 5 14 14
+9 6 11 3 9 4 1 14 4 8
+9 14 12 10 8 11 4 2 11 11
+9 10 4 9 3 12 14 11 15 13
+9 10 10 7 7 10 13 5 14 12
+9 15 7 5 6 2 3 3 9 2
+9 11 8 9 12 4 12 15 5 15
+9 15 2 3 11 1 1 6 6 2
+9 7 10 6 1 8 15 7 9 6
+9 9 6 11 12 4 9 14 12 1
+9 6 3 14 5 6 9 3 2 8
+9 10 6 15 13 2 8 9 11 12
+9 7 12 2 5 15 8 2 14 4
+9 1 6 8 12 13 8 9 6 4
+9 3 3 12 5 2 5 1 11 7
+9 12 7 15 11 2 9 1 3 13
+9 6 9 15 10 8 2 4 5 14
+9 2 5 12 11 2 13 14 5 10
+9 7 7 6 10 14 14 15 14 9
+12 12 2 4 1 6 11 11 12 1 10 2 10
+9 13 2 7 12 8 11 2 5 2
+10 13 7 9 8 9 12 14 9 15 6
+9 15 5 1 4 3 4 2 12 14
+10 3 13 13 12 1 15 6 2 5 2
+20 1 1 3 6 2 2 8 8 7 4 13 14 9 9 1 12 1 4 5 14
+6 6 5 1 3 10 5
+6 8 14 7 15 2 3
+6 14 14 6 5 4 12
+22 14 3 4 2 3 15 11 13 10 5 12 13 11 10 6 10 7 13 12 5 1 10
+6 1 1 13 6 12 12
+6 7 4 8 6 1 10
+6 13 3 3 10 15 8
+6 9 9 12 1 3 9
+6 14 7 7 2 3 14
+27 7 3 7 8 1 10 12 3 8 14 5 15 6 14 14 6 8 11 15 12 12 2 13 3 1 7 5
+30 9 1 6 13 13 1 12 3 4 6 15 9 10 9 11 4 3 3 9 9 2 10 11 3 10 4 14 4 12 4
+30 13 3 5 7 6 4 10 4 4 13 12 2 13 2 2 1 14 8 10 7 7 10 15 14 13 14 3 3 8 3
+30 7 10 7 1 10 9 12 2 8 12 15 4 2 2 11 10 12 14 8 2 11 5 3 9 4 8 9 15 10 4
+30 15 9 3 3 1 9 15 5 1 13 5 6 10 7 14 13 3 15 7 13 13 12 14 1 3 13 2 13 12 7
+20 15 15 10 11 7 11 4 7 12 15 10 11 2 11 12 14 3 6 10 15
+20 13 8 2 10 6 3 11 3 8 1 11 6 13 12 1 14 15 14 11 14
+20 15 7 3 7 4 5 7 1 5 2 13 13 6 4 2 6 11 15 12 4
+20 10 5 14 6 14 3 4 3 4 3 3 5 3 13 7 5 5 12 8 12
+20 9 15 5 9 3 11 9 13 9 11 5 10 3 2 15 5 10 13 2 9
+20 7 6 7 10 15 9 9 15 5 8 11 10 7 14 2 3 7 10 6 11
+20 4 14 7 11 6 15 9 10 7 9 12 6 12 14 13 3 9 12 4 9
+20 9 11 7 8 3 6 12 3 1 13 1 13 1 12 9 15 1 3 13 4
