+++ 2013/cat4/pizza/wedstrijd.uitvoer
@@ -1,0 +1,100 @@
+30
+33
+23
+29
+46
+27
+30
+37
+39
+23
+29
+39
+26
+25
+9
+5
+29
+27
+39
+35
+30
+25
+41
+35
+32
+32
+29
+38
+37
+42
+40
+45
+39
+35
+42
+33
+52
+39
+32
+34
+31
+36
+33
+46
+32
+34
+32
+44
+42
+40
+45
+35
+51
+48
+33
+47
+49
+47
+27
+51
+27
+39
+42
+34
+47
+40
+36
+27
+41
+40
+40
+52
+49
+32
+60
+33
+47
+77
+19
+32
+31
+103
+26
+22
+31
+24
+24
+113
+141
+138
+135
+145
+116
+102
+81
+82
+108
+101
+112
+97
