--- 2013/cat4/pizzakoerier/voorbeeld.invoer
@@ -1,60 +1,0 @@
-9
-6 8
-0 1 8
-1 2 3
-2 3 2
-3 4 4
-4 5 8
-0 5 6
-1 5 5
-2 4 15
-2
-2 4 20 10
-5 4 35 15
-2 1
-0 1 5
-1
-0 1 10 5
-2 1
-0 1 5
-2
-0 1 10 5
-0 1 19 5
-2 1
-0 1 5
-2
-0 1 10 5
-0 1 20 5
-2 1
-0 1 5
-3
-0 1 10 5
-0 1 20 5
-0 1 21 50
-4 4
-0 1 5
-1 2 20
-0 3 5
-2 3 20
-1
-1 2 30 10
-4 4
-0 1 5
-1 2 20
-0 3 5
-2 3 20
-2
-1 2 30 10
-3 2 30 20
-4 4
-0 1 5
-1 2 20
-0 3 5
-2 3 20
-2
-1 2 25 10
-3 2 65 20
-3 1
-1 2 20
-1
-1 2 25 10
