--- 2013/cat4/pizzakoerier/voorbeeld.uitvoer
@@ -1,9 +1,0 @@
-1 15
-2 5
-3 5
-4 10
-5 55
-6 10
-7 20
-8 30
-9 0
