--- 2013/cat4/pizzakoerier/wedstrijd.uitvoer
@@ -1,40 +1,0 @@
-1 15
-2 5
-3 5
-4 10
-5 55
-6 10
-7 20
-8 30
-9 0
-10 0
-11 967
-12 718
-13 940
-14 862
-15 797
-16 949
-17 1085
-18 756
-19 710
-20 977
-21 751
-22 1124
-23 917
-24 1022
-25 814
-26 885
-27 761
-28 1054
-29 1030
-30 1387
-31 1303
-32 1183
-33 1225
-34 1447
-35 1178
-36 1204
-37 989
-38 1173
-39 1182
-40 1116
