+++ 2018/cat4/pizza/cpp/WouterE v2.cpp
