+++ 2018/cat4/pizza/cpp/WouterE.cpp
@@ -1,68 +1,77 @@
 #include <iostream>
 #include <vector>
-#include <map>
 #include <algorithm>
+#include <map>
+#include <set>
 
 using namespace std;
 
-using ii = pair<int, int>;
-using vi = vector<int>;
+struct order { int from, to, time, cost; };
 
-map<ii, int> cache;
+const int inf = 1'000'000'000;
 
-int find_maximum(const vi& pizza, int left, int right) {
-	if ((left + 1) % pizza.size() == right) {
-		return 0;
+vector<vector<int>> graph;
+vector<order> orders;
+map<pair<int, int>, int> cache;
+
+void all_pair_shortest_path() {
+	for (int i = 0; i < graph.size(); i++) {
+		for (int from = 0; from < graph.size(); from++) {
+			for (int to = 0; to < graph.size(); to++) {
+				graph.at(from).at(to) = min(graph.at(from).at(to), graph.at(from).at(i) + graph.at(i).at(to));
+			}
+		}
 	}
-
-	if (cache.find({ left, right }) != cache.end()) {
-		return cache.at({ left, right });
-	}
-
-	int total = 0;
-	for (int i = right; i != (left + 1) % pizza.size(); i = (i + 1) % pizza.size()) {
-		total += pizza.at(i);
-	}
-	
-	int min_other = min(
-		find_maximum(pizza, (left - 1 + pizza.size()) % pizza.size(), right),
-		find_maximum(pizza, left, (right + 1) % pizza.size())
-	);
-
-	cache[{left, right}] = total - min_other;
-	return total - min_other;
 }
 
-int solve(const vi& pizza) {
-	int total = 0;
-	for (int a : pizza) total += a;
+int solve(int start, int time, const set<int>& seen) {
+	if (cache.find({ start, time }) != cache.end()) return cache.at({ start, time });
 
-	int min_bob = total;
-	
-	for (int i = 0; i < pizza.size(); i++) {
-		int max = find_maximum(pizza, (i - 1 + pizza.size()) % pizza.size(), (i + 1) % pizza.size());
-		if (max < min_bob) min_bob = max;
+	int best = 0;
+
+	for (int i = 0; i < orders.size(); i++) {
+		const order& o = orders.at(i);
+
+		if (seen.find(i) == seen.end() && time + graph.at(start).at(o.from) + graph.at(o.from).at(o.to) <= o.time) {
+			set<int> tmp = seen;
+			tmp.insert(i);
+			best = max(best, o.cost + solve(o.to, o.time, tmp));
+		}
 	}
 
-	return total - min_bob;
+	return cache[{start, time}] = best;
 }
 
 int main() {
 	int n;
 	cin >> n;
 
-	for (int i = 0; i < n; i++) {
-		int c;
-		cin >> c;
+	for (int i = 1; i <= n; i++) {
+		int n, m;
+		cin >> n >> m;
 
-		vi pizza;
-		for (int k = 0; k < c; k++) {
-			int slice;
-			cin >> slice;
-			pizza.push_back(slice);
+		graph = vector<vector<int>>(n, vector<int>(n, inf));
+		for (int j = 0; j < n; j++) graph.at(j).at(j) = 0;
+		for (int j = 0; j < m; j++) {
+			int from, to, cost;
+			cin >> from >> to >> cost;
+			graph.at(from).at(to) = cost;
+			graph.at(to).at(from) = cost;
+		}
+
+		all_pair_shortest_path();
+
+		int k;
+		cin >> k;
+
+		orders = vector<order>();
+		for (int j = 0; j < k; j++) {
+			int from, to, time, cost;
+			cin >> from >> to >> time >> cost;
+			orders.push_back({ from, to, time, cost });
 		}
 
 		cache.clear();
-		cout << solve(pizza) << endl;
+		cout << i << " " << solve(0, 0, set<int>()) << endl;
 	}
 }
